requirements.txt
# Python dependencies for Claude development
# Install with: pip install -r requirements.txt

# Anthropic SDK
anthropic>=0.25.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# JSON handling
jsonschema>=4.17.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Code formatting
black>=23.0.0
flake8>=6.0.0

# Type checking
mypy>=1.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Development utilities
ipython>=8.0.0
jupyter>=1.0.0